requirements.txt
fastapi
uvicorn
starlette
pydantic
email-validator
python-multipart
python-dotenv
pymongo
argon2-cffi
python-jose
jinja2
algoliasearch